made class GameOverWorks public
